refactor(widget): move the widget views into KadoCore
The seven widget views, `HabitWidgetCell`, `WidgetMetricsChip` and
`TodayEmptyPlaceholder` become public types in
`KadoCore/Widgets/Views/`; the extension keeps only its `Widget`
configurations, the bundle and the previews. The app can now draw
the same views, which is what the listing's widget screenshot needs —
it is assembled from the widgets rendered in-app rather than
photographed off a Home Screen.

`KadoMark` moves into the package's own asset catalog
(`Image("KadoMark", bundle: .module)`) so the app and the extension
draw one file. Strings stay in the two target catalogs: SwiftUI's
`Text("…")` resolves against `Bundle.main` from a package too, and the
app's catalog already carried the widget keys.

No behaviour change. 566 tests pass.
