+++ KadoWidgets/LockInlineWidget.swift
@@ -2,6 +2,8 @@ import SwiftUI
 import WidgetKit
 import KadoCore
 
+/// The one-line Lock Screen summary. The content is `LockInlineView`,
+/// in `KadoCore`, so the app can draw it too.
 struct LockInlineWidget: Widget {
     let kind: String = "dev.scastiel.kado.widget.lockInline"
 
@@ -13,37 +15,6 @@ struct LockInlineWidget: Widget {
         .configurationDisplayName(Text("Today Summary"))
         .description(Text("One-line summary of today's habits."))
         .supportedFamilies([.accessoryInline])
-    }
-}
-
-struct LockInlineView: View {
-    let entry: SnapshotEntry
-
-    var body: some View {
-        if entry.snapshot.totalDueToday == 0 {
-            Label {
-                Text("No habits due today")
-            } icon: {
-                Image(systemName: "checkmark.circle")
-            }
-        } else {
-            Label {
-                Text(
-                    String(
-                        localized: "\(entry.snapshot.completedToday) of \(entry.snapshot.totalDueToday) done today",
-                        comment: "Inline lock widget summary. Arg 1 completed, arg 2 total."
-                    )
-                )
-            } icon: {
-                Image(systemName: summaryIcon)
-            }
-        }
-    }
-
-    private var summaryIcon: String {
-        if entry.snapshot.totalDueToday == 0 { return "checkmark.circle" }
-        if entry.snapshot.completedToday == entry.snapshot.totalDueToday { return "checkmark.circle.fill" }
-        return "circle.dotted"
     }
 }
