+++ KadoWidgets/TodayProgressMediumWidget.swift
@@ -3,7 +3,8 @@ import WidgetKit
 import KadoCore
 
 /// Medium home widget — two-column habit grid plus a progress
-/// summary. Reads the App Group snapshot.
+/// summary. Reads the App Group snapshot. The content is
+/// `TodayProgressMediumView`, in `KadoCore`, so the app can draw it too.
 struct TodayProgressMediumWidget: Widget {
     let kind: String = "dev.scastiel.kado.widget.todayMedium"
 
@@ -16,62 +17,6 @@ struct TodayProgressMediumWidget: Widget {
         .configurationDisplayName(Text("Today · Progress"))
         .description(Text("Habits due today with a completion summary."))
         .supportedFamilies([.systemMedium])
-    }
-}
-
-struct TodayProgressMediumView: View {
-    let entry: SnapshotEntry
-
-    @Environment(\.widgetRenderingMode) private var renderingMode
-
-    private let limit = 8
-
-    private var palette: WidgetPalette {
-        WidgetPalette(renderingMode: renderingMode)
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            header
-            if entry.snapshot.today.isEmpty {
-                TodayEmptyPlaceholder()
-            } else {
-                cellGrid
-            }
-            Spacer(minLength: 0)
-        }
-    }
-
-    private var header: some View {
-        HStack {
-            Text("Today")
-                .font(.headline)
-                .foregroundStyle(palette.foreground)
-                .widgetAccentable()
-            Spacer()
-            Text(
-                String(
-                    localized: "\(entry.snapshot.completedToday) / \(entry.snapshot.totalDueToday) done",
-                    comment: "Widget progress summary. Arg 1 is completed count, arg 2 is total count."
-                )
-            )
-            .font(.caption.monospacedDigit())
-            .foregroundStyle(palette.foregroundSecondary)
-        }
-    }
-
-    private var cellGrid: some View {
-        LazyVGrid(
-            columns: [
-                GridItem(.flexible(), spacing: 6),
-                GridItem(.flexible(), spacing: 6),
-            ],
-            spacing: 4
-        ) {
-            ForEach(entry.snapshot.today.prefix(limit)) { row in
-                HabitWidgetCell(row: row)
-            }
-        }
     }
 }
