+++ KadoWidgets/WeeklyGridLargeWidget.swift
@@ -3,10 +3,8 @@ import WidgetKit
 import KadoCore
 
 /// Large home widget — habits × last 7 days matrix read from the
-/// App Group snapshot. Layout mirrors the Overview tab: one row
-/// per habit with the name + icon on top and the seven cells
-/// beneath. No horizontal scroll (the widget can't scroll anyway),
-/// so cell width stretches to fill the container.
+/// App Group snapshot. The content is `WeeklyGridLargeView`, in
+/// `KadoCore`, so the app can draw it too.
 struct WeeklyGridLargeWidget: Widget {
     let kind: String = "dev.scastiel.kado.widget.weeklyLarge"
 
@@ -19,197 +17,6 @@ struct WeeklyGridLargeWidget: Widget {
         .configurationDisplayName(Text("This Week"))
         .description(Text("Your habit grid for the past seven days."))
         .supportedFamilies([.systemLarge])
-    }
-}
-
-struct WeeklyGridLargeView: View {
-    let entry: SnapshotEntry
-
-    @Environment(\.widgetRenderingMode) private var renderingMode
-
-    private let rowLimit = 6
-    private static let cellSpacing: CGFloat = 4
-
-    private var palette: WidgetPalette {
-        WidgetPalette(renderingMode: renderingMode)
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Text("This week")
-                .font(.headline)
-                .foregroundStyle(palette.foreground)
-                .widgetAccentable()
-            if entry.snapshot.matrix.isEmpty {
-                emptyPlaceholder
-            } else {
-                weekdayStripe
-                habitRows
-            }
-            Spacer(minLength: 0)
-        }
-    }
-
-    /// Day labels laid out with the same full-width stretch as the
-    /// per-habit cell stripes below, so each label sits directly
-    /// above its column of cells.
-    private var weekdayStripe: some View {
-        GeometryReader { geo in
-            let spacing: CGFloat = Self.cellSpacing
-            let count = CGFloat(max(entry.snapshot.matrixDays.count, 1))
-            let cellWidth = max(
-                8,
-                (geo.size.width - spacing * max(0, count - 1)) / count
-            )
-            HStack(spacing: spacing) {
-                ForEach(entry.snapshot.matrixDays, id: \.self) { day in
-                    Text(weekdayLabel(for: day))
-                        .font(.caption2.monospaced())
-                        .foregroundStyle(isToday(day) ? palette.foreground : palette.foregroundSecondary)
-                        .frame(width: cellWidth)
-                }
-            }
-        }
-        .frame(height: 14)
-    }
-
-    private var habitRows: some View {
-        VStack(alignment: .leading, spacing: 10) {
-            ForEach(entry.snapshot.matrix.prefix(rowLimit), id: \.habit.id) { row in
-                habitBlock(for: row)
-            }
-        }
-    }
-
-    private func habitBlock(for row: WidgetMatrixRow) -> some View {
-        VStack(alignment: .leading, spacing: 4) {
-            // Baseline, not centre: the chip is a step smaller than
-            // the name, and centring two different text sizes floats
-            // the digits above the name's baseline.
-            HStack(alignment: .firstTextBaseline, spacing: 6) {
-                // Only the icon + name are accentable: the metrics
-                // rank second, and content outside the accent group is
-                // already dimmed by the system, which is the whole of
-                // the hierarchy they need.
-                HStack(alignment: .firstTextBaseline, spacing: 6) {
-                    Image(systemName: row.habit.icon)
-                        .font(.caption)
-                        .foregroundStyle(row.habit.color.color)
-                        .frame(width: 16, alignment: .center)
-                    Text(row.habit.name)
-                        .font(.caption.weight(.medium))
-                        .lineLimit(1)
-                        .foregroundStyle(palette.foreground)
-                }
-                .widgetAccentable()
-                Spacer(minLength: 4)
-                WidgetMetricsChip(
-                    streak: row.habit.currentStreak,
-                    scorePercent: row.habit.scorePercent
-                )
-            }
-            cellStripe(for: row)
-        }
-    }
-
-    private func cellStripe(for row: WidgetMatrixRow) -> some View {
-        GeometryReader { geo in
-            let spacing: CGFloat = Self.cellSpacing
-            let count = CGFloat(max(row.cells.count, 1))
-            let cellWidth = max(
-                8,
-                (geo.size.width - spacing * max(0, count - 1)) / count
-            )
-            HStack(spacing: spacing) {
-                ForEach(Array(row.cells.enumerated()), id: \.offset) { _, cell in
-                    WidgetMatrixCell(
-                        cell: cell,
-                        color: row.habit.color,
-                        size: min(cellWidth, 26)
-                    )
-                    .frame(width: cellWidth)
-                }
-            }
-        }
-        .frame(height: 22)
-    }
-
-    private var emptyPlaceholder: some View {
-        VStack(spacing: 6) {
-            Image(systemName: "square.grid.3x3")
-                .font(.title2)
-                .foregroundStyle(palette.foregroundSecondary)
-            Text("No habits yet")
-                .font(.caption)
-                .foregroundStyle(palette.foregroundSecondary)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .widgetAccentable()
-    }
-
-    /// Taken from the snapshot rather than the clock. The app builds
-    /// `matrixDays` ending on the *logical* day, so under a non-midnight
-    /// "Day starts at" hour `isDateInToday` would match no column at all
-    /// between midnight and the rollover.
-    private func isToday(_ day: Date) -> Bool {
-        day == entry.snapshot.matrixDays.last
-    }
-
-    private func weekdayLabel(for day: Date) -> String {
-        let weekdayValue = Calendar.current.component(.weekday, from: day)
-        // `localizedMedium` returns the three-letter standalone
-        // symbol for the current locale ("Sun", "Mon", "dim",
-        // "lun"); truncating to two characters gives the stem the
-        // user expects across languages.
-        return String(Weekday(rawValue: weekdayValue)?.localizedMedium.prefix(2) ?? "")
-    }
-}
-
-/// Widget-local cell view — the app uses `MatrixCell` with a
-/// `DayCell` enum; the widget reads `WidgetDayCell` values
-/// decoded from JSON so we render directly from those.
-struct WidgetMatrixCell: View {
-    let cell: WidgetDayCell
-    let color: HabitColor
-    let size: CGFloat
-
-    @Environment(\.widgetRenderingMode) private var renderingMode
-
-    private var palette: WidgetPalette {
-        WidgetPalette(renderingMode: renderingMode)
-    }
-
-    var body: some View {
-        RoundedRectangle(cornerRadius: 4, style: .continuous)
-            .fill(fill)
-            .overlay {
-                // Hollow = logged off-schedule. Same language the
-                // main app's `MatrixCell` uses, thinner border for
-                // the smaller cell.
-                if let borderOpacity = cell.borderOpacity {
-                    RoundedRectangle(cornerRadius: 4, style: .continuous)
-                        .strokeBorder(
-                            color.color.opacity(borderOpacity),
-                            lineWidth: 1.5
-                        )
-                }
-            }
-            .frame(height: size)
-    }
-
-    /// `.scored` / `.offSchedule` already carry their own alpha, so
-    /// they survive the tint untouched. `.notDue` is the one opaque
-    /// fill here, and an opaque fill is exactly what gets flattened
-    /// into a solid block under Tinted / Clear — route it through the
-    /// palette, which keeps it under the scored ramp's 0.2 floor so
-    /// "never due" stays tellable from "due and missed".
-    private var fill: Color {
-        switch cell {
-        case .future: Color.clear
-        case .notDue: palette.notDueFill
-        case .scored: color.color.opacity(cell.colorOpacity ?? 0)
-        case .offSchedule: color.color.opacity(cell.offScheduleFillOpacity ?? 0)
-        }
     }
 }
