+++ Packages/KadoCore/Sources/KadoCore/Resources/Media.xcassets/KadoMark.imageset/Contents.json
