+++ Packages/KadoCore/Sources/KadoCore/Resources/Media.xcassets/KadoMark.imageset/kado-mark@2x.png
