+++ Packages/KadoCore/Sources/KadoCore/Resources/Media.xcassets/KadoMark.imageset/kado-mark@3x.png
